[NO JIRA] Amazon changed two page signin to one page
