+++ .gitignore
@@ -12,7 +12,6 @@ tmp/
 *~.nib
 local.properties
 .settings/
-.target/
 .loadpath
 .recommenders
